+++ app.py
@@ -1,24 +1,7 @@
-import os
 from config import CONFIG
-from functools import wraps
-from flask import Flask, redirect, abort, request
+from flask import Flask, redirect
 
 app = Flask(__name__)
-
-
-def auth(view):
-    @wraps(view)
-    def check_pw(*args, **kwargs):
-        # TODO: this is vulnerable to a timing attack
-        password = request.args.get("password")
-        PW = os.environ["PRINTER_PW"] if "PRINTER_PW" in os.environ else None
-        if password != PW:
-            abort(401)
-        else:
-            return view(*args, **kwargs)
-
-    return check_pw
-
 
 if CONFIG.printer:
     from mods.printer import printer
